Fix Moscow timezone.
DST changes started in 1981 but used several different patterns.

Moscow briefly used EET/EEST timezone in 1991, but switched back on
January 19, 1992.  Since this file format cannot express this, just
ignore it.  MSK/MSD just went on as usual, it's just Moscow was not
actually in it.

It moved to GMT+4 without DST changes in 2011.  Technically this
should be called "MSK" (with a different offset), but just treat it as
permanent DST (which it was) and let it be called "MSD". (Again, the
file format cannot express this properly).

It moved back to GMT+3 without DST changes in 2014.

FossilOrigin-Name: 06047a700343e453689585b578c934eacf70313043381bcec01d337eb22f057d